Rotas Arrumadas
- helpers.js movidos para ./Components;
- FormPage antiga deixada de lado (old);
- Variável password mudada pra senha na LoginPage;
