+++ client/src/app/index.jsx
@@ -3,7 +3,7 @@ import { Button, StyleSheet, View } from 'react-native';
 import { useRouter } from 'expo-router'
 
 // Components
-import { LoginButton, CreateAccountButton, DevelopmentButton } from '../Components/Buttons';
+import { LoginButton, DevelopmentButton } from '../Components/Buttons';
 
 export default function Home() {
   const router = useRouter();
